backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
python-multipart==0.0.6
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
langchain==0.1.0
pymupdf==1.23.7
python-dotenv==1.0.0
pydantic==1.10.8
faiss-cpu==1.7.4
tiktoken==0.5.1
huggingface_hub==0.20.3
sentence-transformers==2.2.2
langchain-community==0.0.13
langsmith>=0.0.77,<0.1.0